+++ inStrain/profile/samtools_ops.py
@@ -8,9 +8,9 @@ import pysam
 import logging
 from subprocess import call
 from Bio import bgzf
+from subprocess import PIPE, run, Popen
 
-
-def prepare_bam_fie(bam, processes=6):
+def prepare_bam_fie(bam, processes=6, max_reads=None):
     """
     Make this a .bam file
 
@@ -35,9 +35,21 @@ def prepare_bam_fie(bam, processes=6):
         # If sorted, just index
         else:
             logging.info("Indexing {0}".format(bam))
-            cmd = ['samtools', 'index', bam, bam + '.bai', '-@', str(processes)]
-            print(' '.join(cmd))
-            call(cmd)
+            _index_bam(bam, processes=processes)
+
+    if max_reads is not None:
+        total = calc_number_reads(bam)
+        ratio = int(max_reads) / total
+
+        if ratio < 1:
+            seed = str(42) + f"{ratio:.3f}"[1:]
+
+            new_bam = bam + '.sub'
+
+            logging.info(f"Subsetting with seed {seed}")
+            bam = subset_bam(bam, seed, new_bam, p=processes)
+
+            _index_bam(bam, processes=processes)
 
     if os.stat(bam).st_size == 0:
         logging.error("Failed to generated a sorted .bam file! Make sure you have " +
@@ -71,6 +83,18 @@ def _sam_to_bam(sam, processes=6):
 
     return bam
 
+def _index_bam(bam, processes=6):
+    # Try standard way
+    cmd = ['samtools', 'index', bam, bam + '.bai', '-@', str(processes)]
+    print(' '.join(cmd))
+    call(cmd)
+    exit_code = call(cmd)
+
+    if exit_code != 0:
+        # Try non-standard way
+        cmd = ['samtools',  'index', '-@', str(processes), bam]
+        print(' '.join(cmd))
+        call(cmd)
 
 def _sort_index_bam(bam, processes=6, rm_ori=False):
     """
@@ -87,14 +111,14 @@ def _sort_index_bam(bam, processes=6, rm_ori=False):
         cmd = ['samtools', 'sort', bam, '-o', sorted_bam, '-@', str(processes)]
         print(' '.join(cmd))
         call(cmd)
+
     else:
         sorted_bam = bam
         rm_ori = False
 
     logging.info("Indexing {0}".format(sorted_bam))
-    cmd = ['samtools', 'index', sorted_bam, sorted_bam + '.bai', '-@', str(processes)]
-    print(' '.join(cmd))
-    call(cmd)
+    _index_bam(sorted_bam, processes=processes)
+
 
     if rm_ori:
         logging.info("Deleting {0}".format(bam))
@@ -110,3 +134,31 @@ def _is_sorted_bam(bam):
     with bgzf.BgzfReader(bam, "rb") as fin:
         bam_header = fin.readline().strip()
         return b"SO:coordinate" in bam_header
+
+def calc_number_reads(bam):
+    """
+    Return the number of reads in the .bam file
+    """
+    cmd = ['samtools', 'idxstats', bam]
+    logging.info(' '.join(cmd))
+
+    # run idxstats
+    result = run(cmd, stdout=PIPE, stderr=PIPE, universal_newlines=True)
+    assert result.returncode == 0, result
+
+    # sum number of reads
+    total_reads = 0
+    for line in result.stdout.split('\n'):
+        if len(line.split()) != 4:
+            continue
+        total_reads += (int(line.split()[2]) + int(line.split()[3]))
+
+    return total_reads
+
+def subset_bam(ori_bam, substr, new_bam, p=6):
+    cmd = f"samtools view -@ {p} -bs {substr} {ori_bam} > {new_bam}"
+    #cmd = f"sambamba view -s {substr} -t {p} -f bam {ori_bam} > {new_bam}"
+    logging.info(cmd)
+    result = run(cmd, stdout=PIPE, stderr=PIPE, universal_newlines=True, shell=True)
+    assert result.returncode == 0, result
+    return new_bam