+++ test/test_suite.py
@@ -20,7 +20,7 @@ import tests.test_special
 warnings.filterwarnings("ignore")
 
 pd.set_option('display.max_columns', None)
-pd.set_option('display.max_colwidth', -1)
+pd.set_option('display.max_colwidth', None)
 
 
 if __name__ == '__main__':